[fixed] improvements on setAppElement...
There was some pitfalls on how `setAppElement` work.

If your <script /> was in <head />, there was a change that it tries
to use `document.body` that is not yet ready.

Another one was using an selector string that does not find any
elements, causing it to try to perform all call on `null`.

This patch can also help if you want to do server-side rendering,
but this was not tested and, perhaps, it's better to use this function
correctly.